Change Folder Structure, add SQL
I opened an issue about this explaining the idea behind this folder
structure change, and have branched it so that it won't break what
you're working on. Let me know if this works for you and then we can
merge it with master.

In addition, I added a database folder for any .sql scripts so you can
create all of the necessary database objects on your local server.